+++ packages/vue/__tests__/Mixin/Collection.test.js
@@ -53,6 +53,9 @@ describe('The Collection mixin', () => {
       collectionName: 'products',
       $store: {
         commit: jest.fn(),
+        getters: {
+          'hyral_product/collection': jest.fn(() => null),
+        },
       },
       initCollection: collectionMixin.methods.initCollection,
       loadCollection: collectionMixin.methods.loadCollection,
@@ -86,6 +89,31 @@ describe('The Collection mixin', () => {
     });
   });
 
+  test('that the initCollection correctly initializes the store', () => {
+    const state = {
+      collections: {},
+    };
+    const mixin = Object.assign({
+      resourceType: 'product',
+      collectionName: 'products',
+      $store: {
+        commit: jest.fn((s, collection) => { state.collections.products = collection }),
+        getters: {
+          'hyral_product/collection': jest.fn(() => state.collections.products || null),
+        },
+      },
+      initCollection: collectionMixin.methods.initCollection,
+    }, collectionMixin);
+
+    mixin.initCollection();
+
+    expect(mixin.$store.commit).toHaveBeenCalledTimes(1);
+
+    mixin.initCollection();
+
+    expect(mixin.$store.commit).toHaveBeenCalledTimes(1);
+  });
+
   test('that the mixin handles errors on mounted', () => {
     const mixin = Object.assign({
       $store: {
@@ -98,5 +126,24 @@ describe('The Collection mixin', () => {
     mixin.mounted.call(mixin);
 
     expect(mixin.$store.getters['hyral_product/collection']).not.toHaveBeenCalled();
+  });
+
+  test('that the mixin doesn\'t initialize if the component does not define the collectionName or resourceType', () => {
+    const mixin = Object.assign({
+      initCollection: jest.fn(),
+      loadCollection: jest.fn(),
+    }, collectionMixin);
+
+    mixin.serverPrefetch().catch(() => {});
+    expect(mixin.initCollection).not.toHaveBeenCalled();
+    expect(mixin.loadCollection).not.toHaveBeenCalled();
+
+    mixin.mounted();
+    expect(mixin.initCollection).not.toHaveBeenCalled();
+    expect(mixin.loadCollection).not.toHaveBeenCalled();
+
+    mixin.created();
+    expect(mixin.initCollection).not.toHaveBeenCalled();
+    expect(mixin.loadCollection).not.toHaveBeenCalled();
   });
 });