+++ packages/core/CHANGELOG.md
@@ -3,6 +3,37 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.3.0](https://github.com/SyneticNL/Hyral/compare/v1.2.0...v1.3.0) (2019-09-17)
+
+
+### Features
+
+* **core:** allow responseType to be defined by the responseNormalizer ([6e72ab5](https://github.com/SyneticNL/Hyral/commit/6e72ab5))
+
+
+
+
+
+# [1.2.0](https://github.com/SyneticNL/Hyral/compare/v1.1.1...v1.2.0) (2019-08-30)
+
+
+### Bug Fixes
+
+* **core:** fromState returns incorrect relationship types for resources ([4193e49](https://github.com/SyneticNL/Hyral/commit/4193e49))
+
+
+### Features
+
+* **core:** add support for ParameterBag in Repository.findById [#82](https://github.com/SyneticNL/Hyral/issues/82) ([9b2e79e](https://github.com/SyneticNL/Hyral/commit/9b2e79e))
+* **core:** delete support in persistCascade for 1-to-1 relations [#42](https://github.com/SyneticNL/Hyral/issues/42) ([03e0281](https://github.com/SyneticNL/Hyral/commit/03e0281))
+* **core:** delete support in persistCascade for 1-to-1 relations [#42](https://github.com/SyneticNL/Hyral/issues/42) ([e9aca6f](https://github.com/SyneticNL/Hyral/commit/e9aca6f))
+* **vue:** add support for ParameterBag in createStoreModule [#82](https://github.com/SyneticNL/Hyral/issues/82) ([e4f7750](https://github.com/SyneticNL/Hyral/commit/e4f7750))
+* **vue:** add support for ParameterBag in resource mixin [#82](https://github.com/SyneticNL/Hyral/issues/82) ([fc0477b](https://github.com/SyneticNL/Hyral/commit/fc0477b))
+
+
+
+
+
 ## [1.1.1](https://github.com/SyneticNL/Hyral/compare/v1.1.0...v1.1.1) (2019-08-02)